requirements.txt
flask
requests
Werkzeug
Jinja2
pandas
google-cloud-translate
vertexai 
